requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
arrow-cpp=0.13.0=py37h117bdfb_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
boost-cpp=1.67.0=h14c3975_4
brotli=1.0.7=he6710b0_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.11.27=0
certifi=2019.11.28=py37_0
cycler=0.10.0=py37_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py37_1
double-conversion=3.1.5=he6710b0_1
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gflags=2.2.2=he6710b0_0
gitdb2=2.0.6=py_0
gitpython=3.0.5=py_0
glib=2.56.2=hd408876_0
glog=0.4.0=he6710b0_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
importlib_metadata=1.1.0=py37_0
intel-openmp=2019.4=243
ipykernel=5.1.2=py37h39e3cac_0
ipython=7.8.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
jedi=0.15.1=py37_0
joblib=0.14.0=py_0
jpeg=9b=h024ee3a_2
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.0=py37_0
kiwisolver=1.1.0=py37he6710b0_0
libboost=1.67.0=h46d08c1_4
libedit=3.1.20181209=hc058e9b_0
libevent=2.1.8=h1ba5d50_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.6.0=hdbcaa40_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
line_profiler=2.1.2=py37h14c3975_0
llvmlite=0.30.0=py37hd408876_0
lz4-c=1.8.1.2=h14c3975_0
matplotlib=3.1.1=py37h5429711_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.14=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
more-itertools=7.2.0=py37_0
ncurses=6.1=he6710b0_1
numba=0.46.0=py37h962f231_0
numpy=1.17.2=py37haad9e8e_0
numpy-base=1.17.2=py37hde5b4d6_0
openssl=1.1.1d=h7b6447c_3
packaging=19.2=py_0
pandas=0.25.1=py37he6710b0_0
parso=0.5.1=py_0
patsy=0.5.1=py37_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pip=19.2.3=py37_0
pluggy=0.13.1=py37_0
prompt_toolkit=2.0.10=py_0
ptyprocess=0.6.0=py37_0
py=1.8.0=py37_0
pyarrow=0.13.0=py37he6710b0_0
pygments=2.4.2=py_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37h05f1152_2
pytest=5.3.1=py37_0
python=3.7.4=h265db76_1
python-dateutil=2.8.0=py37_0
pytz=2019.3=py_0
pyzmq=18.1.0=py37he6710b0_0
qt=5.9.7=h5867ecd_1
re2=2019.08.01=he6710b0_0
readline=7.0=h7b6447c_5
scipy=1.3.1=py37h7c811a0_0
seaborn=0.9.0=pyh91ea838_1
setuptools=41.4.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=py37_0
smmap2=2.0.5=py37_0
snappy=1.1.7=hbae5bb6_3
sqlite=3.30.0=h7b6447c_0
statsmodels=0.10.1=py37hdd07704_0
tbb=2019.8=hfd86e86_0
thrift-cpp=0.11.0=h02b749d_3
tk=8.6.8=hbc83047_0
tornado=6.0.3=py37h7b6447c_0
traitlets=4.3.3=py37_0
wcwidth=0.1.7=py37_0
wheel=0.33.6=py37_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
